requirements.txt
claude-agent-sdk
anthropic
httpx
python-dotenv
fastapi
uvicorn[standard]
supabase
PyJWT
pandas
psycopg2-binary
